+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFZQC13Z8MN14Y4RRMKB1YTD"
+        "01KG1DFQH9PXQ2KVWY2JSRHJDB"
     )
     r"""The unique identifier of the data source"""