+++ packages/orq-rc/src/orq_ai_sdk/models/updateconversationop.py
@@ -231,7 +231,7 @@ class UpdateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kfzqc0m4gtsn5bjztwsvfvtw"
+        "conv_01kg1dfq284cdsxdq652v8qg71"
     )
     r"""Unique conversation identifier with `conv_` prefix."""