+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFZQC149N2ET1NXSANJKD241"
+        "01KG1DFQHC88C1CTKJ5FX36EKW"
     )
     r"""The unique identifier of the data source"""